requirements.txt
google-cloud-aiplatform
google-cloud-storage
google-cloud-logging
google-adk
google-adk[eval]
requests
gcsfs
pypdf
matplotlib
numpy
pandas
scikit-learn
fsspec
tabulate
PyYAML
opentelemetry-exporter-gcp-trace
fastapi
uvicorn
pydantic
python-dotenv
google-cloud-discoveryengine

